move hash generation to base class
