+++ contracts/randomizer/BasicRandomizer.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 // Creatd By: Art Blocks Inc.
 
-import "./interfaces/0.8.x/IRandomizer.sol";
+import "../interfaces/0.8.x/IRandomizer.sol";
 
 pragma solidity 0.8.9;
